+++ apps/agentic-server/package.json
@@ -19,6 +19,10 @@
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
+    "mastra": "^0.10.3",
+    "@mastra/core": "^0.10.3",
+    "@mastra/memory": "^0.10.2",
+    "@mastra/libsql": "^0.10.1",
     "zod": "^3.25.32"
   }
 }
