+++ apps/agentic-server/src/mastra/index.ts
@@ -1,0 +1,16 @@
+import { Mastra } from '@mastra/core/mastra';
+import { PinoLogger } from '@mastra/loggers';
+import { LibSQLStore } from '@mastra/libsql';
+import { shapeshiftAgent } from './agents/shapeshift-agent';
+
+export const mastra = new Mastra({
+  agents: { shapeshiftAgent },
+  storage: new LibSQLStore({
+    // stores telemetry, evals, ... into memory storage, if it needs to persist, change to file:../mastra.db
+    url: ':memory:',
+  }),
+  logger: new PinoLogger({
+    name: 'Mastra',
+    level: 'info',
+  }),
+});
