+++ caip/src/lib/assetId.ts
@@ -120,7 +120,8 @@ export const toAssetId: ToAssetId = (args: ToAssetIdArgs): AssetId => {
       case 'erc721':
       case 'erc1155':
       case 'bep721':
-      case 'bep1155': { // caip-29 (https://github.com/ChainAgnostic/CAIPs/blob/master/CAIPs/caip-29.md) // caip-22 (https://github.com/ChainAgnostic/CAIPs/blob/master/CAIPs/caip-22.md)
+      case 'bep1155': {
+        // caip-29 (https://github.com/ChainAgnostic/CAIPs/blob/master/CAIPs/caip-29.md) // caip-22 (https://github.com/ChainAgnostic/CAIPs/blob/master/CAIPs/caip-22.md)
         const [address] = assetReference.split('/');
         assertContractAddress(address);
         return assetReference.toLowerCase();
